test: split PDF source geometry cases
